docs(chess): split modules, apply EXAMPLE_STANDARD, pin cougr-core 1.1.0
Splits the monolithic lib.rs into components.rs (component types),
systems.rs (pure move/board logic), and zk.rs (Groth16 circuit
submission, relocated unchanged) per EXAMPLE_STANDARD.md §3. Rewrites
README.md to cover all eight §4 sections, marks the example
transitional, bumps cougr-core to 1.1.0, and commits Cargo.lock. No
game rule or ZK verification logic was changed — this is out of
scope per the parent issue.

Refs #223
